pyproject.toml
[project]
name = "music-player"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "python-vlc>=3.0.21203",
    "yt-dlp>=2026.3.17",
    "ytmusicapi>=1.11.5",
]
